remove dpr multiple from selection item/svg context
